requirements_versions.txt
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
torchtext==0.16.0
xformers==0.0.22.post7
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
gradio==3.50.2
einops==0.7.0
transformers==4.35.2
accelerate==0.24.1
safetensors==0.4.0
scipy==1.11.3
tqdm==4.66.1
psutil==5.9.6
pytorch_lightning==2.1.2
omegaconf==2.3.0
pytorch3d==0.7.6
moviepy==1.0.3
imageio==2.31.1
imageio-ffmpeg==0.4.9